Move streamconfig/initializers.go to consumer/producer
